+++ packages/auto_mappr/lib/src/extensions/dart_object_extension.dart
@@ -1,7 +1,9 @@
 //ignore_for_file: no-object-declaration
 
 import 'package:analyzer/dart/constant/value.dart';
+import 'package:auto_mappr/src/extensions/element_extension.dart';
 import 'package:auto_mappr/src/extensions/executable_element_extension.dart';
+import 'package:auto_mappr/src/models/auto_mappr_config.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -10,6 +12,7 @@ extension DartObjectExtension on DartObject {
   ///
   /// Otherwise return code expression of literals or objects.
   Expression? toCodeExpression({
+    required AutoMapprConfig config,
     bool passModelArgument = false,
   }) {
     if (isNull) {
@@ -25,24 +28,34 @@ extension DartObjectExtension on DartObject {
     }
 
     final emitter = DartEmitter();
-    final output = _toSpec().accept(emitter);
+    final output = _ToCodeExpressionConverter(config: config).convert(this).accept(emitter);
 
     return CodeExpression(Code('$output'));
   }
+}
 
-  Spec _toSpec() {
-    final constant = ConstantReader(this);
+class _ToCodeExpressionConverter {
+  final AutoMapprConfig config;
+
+  _ToCodeExpressionConverter({required this.config});
+
+  Spec convert(DartObject dartObject) {
+    return _toSpec(dartObject);
+  }
+
+  Spec _toSpec(DartObject dartObject) {
+    final constant = ConstantReader(dartObject);
 
     if (constant.isLiteral) {
-      return _reviveLiteral();
+      return _reviveLiteral(dartObject);
     }
 
     /// Perhaps an object instantiation?
-    return _reviveObject();
+    return _reviveObject(dartObject);
   }
 
-  Spec _reviveLiteral() {
-    final constant = ConstantReader(this);
+  Spec _reviveLiteral(DartObject dartObject) {
+    final constant = ConstantReader(dartObject);
 
     // Special literals
 
@@ -61,47 +74,50 @@ extension DartObjectExtension on DartObject {
     // Collections
 
     if (constant.isList) {
-      return literalList(constant.listValue.map((item) => item._toSpec()));
+      return literalList(constant.listValue.map(_toSpec));
     }
 
     if (constant.isSet) {
-      return literalSet(constant.setValue.map((item) => item._toSpec()));
+      return literalSet(constant.setValue.map(_toSpec));
     }
 
     if (constant.isMap) {
-      return literalMap(constant.mapValue.map((key, value) => MapEntry(key!._toSpec(), value!._toSpec())));
+      return literalMap(
+        constant.mapValue.map((key, value) => MapEntry(_toSpec(key!), _toSpec(value!))),
+      );
     }
 
     // int, double, num, null
     return literal(constant.literalValue);
   }
 
-  Spec _reviveObject() {
-    final revived = ConstantReader(this).revive();
+  Spec _reviveObject(DartObject dartObject) {
+    final revived = ConstantReader(dartObject).revive();
 
     final location = revived.source.toString().split('#');
+    final libraryAlias = dartObject.type!.element!.getLibraryAlias(config: config);
 
-    /// Getters, Setters, Methods can't be declared as constants so this
-    /// literal must either be a top-level constant or a static constant and
-    /// can be directly accessed by `revived.accessor`.
+    // Getters, Setters, Methods can't be declared as constants so this
+    // literal must either be a top-level constant or a static constant and
+    // can be directly accessed by `revived.accessor`.
     if (location.length <= 1) {
-      return refer(revived.accessor);
+      return refer('$libraryAlias${revived.accessor}');
     }
 
-    /// If this is a class instantiation then `location[1]` will be populated
-    /// with the class name.
-    ///
-    /// location[1] - class
-    /// revived.accessor - named constructor
+    // If this is a class instantiation then `location[1]` will be populated
+    // with the class name.
+    //
+    // location[1] - class
+    // revived.accessor - named constructor
     final instantiation = location[1];
     final useNamedConstructor = revived.accessor.isNotEmpty;
 
-    final revivedInstance = refer(instantiation);
+    final revivedInstance = refer('$libraryAlias$instantiation');
 
     final positionalArguments =
-        revived.positionalArguments.map<Expression>((argument) => argument._toSpec() as Expression);
+        revived.positionalArguments.map<Expression>((argument) => _toSpec(argument) as Expression);
     final namedArguments =
-        revived.namedArguments.map<String, Expression>((key, value) => MapEntry(key, value._toSpec() as Expression));
+        revived.namedArguments.map<String, Expression>((key, value) => MapEntry(key, _toSpec(value) as Expression));
 
     if (useNamedConstructor) {
       return revivedInstance.constInstanceNamed(revived.accessor, positionalArguments, namedArguments);