+++ packages/auto_mappr/lib/src/extensions/element_extension.dart
@@ -2,20 +2,19 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:auto_mappr/src/models/auto_mappr_config.dart';
 
 extension ElementExtension on Element {
+  /// Returns an library alias with [postfix] (usually '.'),
+  /// or empty string if no alias detected.
   String getLibraryAlias({
     required AutoMapprConfig config,
+    String postfix = '.',
   }) {
-    var alias = '';
+    final libraryUri = source?.uri.toString();
+    final contains = config.libraryUriToAlias.containsKey(libraryUri);
 
-    if (!(library?.isInSdk ?? false)) {
-      final libraryUri = source?.uri.toString();
-      final contains = config.libraryUriToAlias.containsKey(libraryUri);
-
-      if (contains) {
-        alias = config.libraryUriToAlias[libraryUri]!;
-      }
+    if (contains) {
+      return '${config.libraryUriToAlias[libraryUri]!}$postfix';
     }
 
-    return alias;
+    return '';
   }
 }