+++ packages/auto_mappr/CHANGELOG.md
@@ -1,7 +1,8 @@
 [//]: # (## Unreleased)
 
 ## Unreleased
-- Add support for library aliases, so mapping now supports types from different libraries with the same name. [#62](https://github.com/netglade/auto_mappr/pull/62)
+- Add support for library aliases, so mapping now supports types from different libraries with the same name.
+It also work in cases when library alias's exports. [#62](https://github.com/netglade/auto_mappr/pull/62)
 
 ## 1.4.0
 - Add modules. [#67](https://github.com/netglade/auto_mappr/pull/67)